# src/backend/openapi.yaml
openapi: 3.0.3
info:
  title: Health & Wellbeing E-Commerce API
  description: |
    Comprehensive API documentation for the Health and Wellbeing E-Commerce platform.
    
    ## Features
    * Product catalog with filtering and search
    * Category and tag management
    * Guest and authenticated checkout
    * User authentication and authorization
    * Admin product and order management
    
    ## Authentication
    Most endpoints are public. Admin endpoints require JWT token with ADMIN role.
    
    To authenticate, use the `/api/auth/login` endpoint and include the returned token
    in the Authorization header: `Bearer <token>`
  version: 1.0.0
  contact:
    name: API Support
    email: support@welbeing.com
  license:
    name: MIT

servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.welbeing.com
    description: Production server

tags:
  - name: Authentication
    description: User registration and login
  - name: Products
    description: Product catalog operations
  - name: Categories
    description: Product category management
  - name: Tags
    description: Wellbeing tag management
  - name: Orders
    description: Order management for users
  - name: Admin Products
    description: Admin-only product management
  - name: Admin Orders
    description: Admin-only order management

paths:
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Create a new user account with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      description: Authenticate user with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/products:
    get:
      tags:
        - Products
      summary: List products
      description: Retrieve a list of products with optional filtering
      parameters:
        - name: category
          in: query
          description: Filter by category ID
          schema:
            type: integer
        - name: tag
          in: query
          description: Filter by tag ID
          schema:
            type: integer
        - name: search
          in: query
          description: Search in product name and description
          schema:
            type: string
      responses:
        '200':
          description: Products retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductsResponse'

  /api/products/{id}:
    get:
      tags:
        - Products
      summary: Get product details
      description: Retrieve detailed information about a specific product
      parameters:
        - name: id
          in: path
          required: true
          description: Product ID
          schema:
            type: integer
      responses:
        '200':
          description: Product retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/categories:
    get:
      tags:
        - Categories
      summary: List categories
      description: Retrieve all product categories
      responses:
        '200':
          description: Categories retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoriesResponse'

  /api/tags:
    get:
      tags:
        - Tags
      summary: List wellbeing tags
      description: Retrieve all wellbeing tags
      responses:
        '200':
          description: Tags retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagsResponse'

  /api/orders:
    get:
      tags:
        - Orders
      summary: Get user orders
      description: Retrieve orders for the authenticated user
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Orders retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrdersResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - Orders
      summary: Create order
      description: Create a new order (guest or authenticated). Provide address_id for saved addresses or shipping_address for guest checkout.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Invalid order data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/admin/products:
    get:
      tags:
        - Admin Products
      summary: List all products (Admin)
      description: Retrieve all products including out of stock and hidden ones
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Products retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductsResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - Admin access only
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - Admin Products
      summary: Create product (Admin)
      description: Create a new product with multiple images and tags
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Invalid product data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - Admin access only
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/admin/products/{id}:
    put:
      tags:
        - Admin Products
      summary: Update product (Admin)
      description: Update an existing product
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Product ID
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProductRequest'
      responses:
        '200':
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Invalid product data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - Admin access only
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - Admin Products
      summary: Delete product (Admin)
      description: Delete a product (cannot delete products with existing orders)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Product ID
          schema:
            type: integer
      responses:
        '204':
          description: Product deleted successfully
        '400':
          description: Cannot delete product with orders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - Admin access only
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/admin/orders:
    get:
      tags:
        - Admin Orders
      summary: List all orders (Admin)
      description: Retrieve all orders in the system
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Orders retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrdersResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - Admin access only
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/admin/orders/{id}/status:
    patch:
      tags:
        - Admin Orders
      summary: Update order status (Admin)
      description: Update the status of an order
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Order ID
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum: [PENDING, PAID, SHIPPED, DELIVERED, CANCELLED]
      responses:
        '200':
          description: Order status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Invalid status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - Admin access only
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/auth/login

  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          format: email
          example: user@example.com
        firstName:
          type: string
          example: Jordan
        lastName:
          type: string
          example: Lee
        phone:
          type: string
          nullable: true
          example: "+15550000002"
        role:
          type: string
          enum: [USER, ADMIN]
          example: USER
        isActive:
          type: boolean
          example: true
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
          example: '2026-01-10T09:00:00Z'

    AuthResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT token for authentication
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        user:
          $ref: '#/components/schemas/User'

    RegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          minLength: 6
          example: securepassword123
        firstName:
          type: string
          description: Optional first name (defaults to "User")
          example: Jordan
        lastName:
          type: string
          description: Optional last name (defaults to "User")
          example: Lee

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          example: securepassword123

    ProductImage:
      type: object
      properties:
        id:
          type: integer
          example: 1
        url:
          type: string
          format: uri
          example: https://example.com/images/product1.jpg
        displayOrder:
          type: integer
          example: 0
        altText:
          type: string
          nullable: true
          example: Knee support image

    ProductVariant:
      type: object
      properties:
        id:
          type: integer
          example: 10
        sku:
          type: string
          example: SUP-KNEE-001-BLK-M
        price:
          type: number
          format: decimal
          example: 29.99
        stockQuantity:
          type: integer
          example: 3
        optionValues:
          type: object
          additionalProperties:
            type: string
          example:
            size: M
            color: Black
        isActive:
          type: boolean
          example: true

    WellbeingTag:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Joint Pain
        type:
          type: string
          enum: [GOAL, FEATURE, NEED]
          example: GOAL

    Category:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Supplements
        description:
          type: string
          nullable: true
          example: Vitamins and minerals

    Product:
      type: object
      properties:
        id:
          type: integer
          example: 1
        sku:
          type: string
          nullable: true
          example: SUP-KNEE-001
        name:
          type: string
          example: Knee Support Pro
        description:
          type: string
          example: Advanced knee support for active lifestyles
        price:
          type: number
          format: decimal
          example: 29.99
        originalPrice:
          type: number
          format: decimal
          nullable: true
          example: 39.99
        stockQuantity:
          type: integer
          example: 10
        reorderLevel:
          type: integer
          example: 5
        isVisible:
          type: boolean
          example: true
        ingredients:
          type: string
          nullable: true
        usageInstructions:
          type: string
          nullable: true
        benefits:
          type: string
          nullable: true
        safetyDisclaimers:
          type: string
          nullable: true
        categoryId:
          type: integer
          example: 1
        category:
          $ref: '#/components/schemas/Category'
        images:
          type: array
          items:
            $ref: '#/components/schemas/ProductImage'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/WellbeingTag'
        variants:
          type: array
          items:
            $ref: '#/components/schemas/ProductVariant'

    ProductResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Product'

    ProductsResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/Product'

    CategoriesResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/Category'

    TagsResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/WellbeingTag'

    CreateProductRequest:
      type: object
      required:
        - name
        - description
        - price
        - categoryId
      properties:
        sku:
          type: string
          nullable: true
          example: SUP-KNEE-001
        name:
          type: string
          example: Knee Support Pro
        description:
          type: string
          example: Advanced knee support for active lifestyles
        price:
          type: number
          format: decimal
          example: 29.99
        originalPrice:
          type: number
          format: decimal
          nullable: true
          example: 39.99
        reorderLevel:
          type: integer
          default: 0
        categoryId:
          type: integer
          example: 1
        stockQuantity:
          type: integer
          default: 0
        isVisible:
          type: boolean
          default: true
        ingredients:
          type: string
          nullable: true
        usageInstructions:
          type: string
          nullable: true
        benefits:
          type: string
          nullable: true
        safetyDisclaimers:
          type: string
          nullable: true
        imageUrls:
          type: array
          items:
            type: string
          description: Array of image URLs (one per line in admin form)
          example:
            - https://example.com/image1.jpg
            - https://example.com/image2.jpg
        tagIds:
          type: array
          items:
            type: integer
          description: Array of tag IDs to associate
          example:
            - 1
            - 2

    UpdateProductRequest:
      allOf:
        - $ref: '#/components/schemas/CreateProductRequest'
      type: object
      description: All fields are optional for partial updates

    OrderItem:
      type: object
      properties:
        id:
          type: integer
          example: 1
        quantity:
          type: integer
          example: 2
        priceAtPurchase:
          type: number
          format: decimal
          example: 59.98
        productVariant:
          $ref: '#/components/schemas/ProductVariant'
        product:
          $ref: '#/components/schemas/Product'

    Address:
      type: object
      properties:
        id:
          type: integer
          example: 3
        label:
          type: string
          example: Home
        fullName:
          type: string
          example: Jordan Lee
        phone:
          type: string
          example: "+15550000002"
        streetLine1:
          type: string
          example: 123 Wellness Ave
        streetLine2:
          type: string
          nullable: true
          example: Apt 4B
        city:
          type: string
          example: Austin
        state:
          type: string
          example: TX
        postalCode:
          type: string
          example: 73301
        country:
          type: string
          example: US
        isDefault:
          type: boolean
          example: true

    Payment:
      type: object
      properties:
        id:
          type: integer
          example: 1
        provider:
          type: string
          enum: [STRIPE, PAYPAL, COD, MOCK]
          example: STRIPE
        transactionId:
          type: string
          example: pi_123
        amount:
          type: number
          format: decimal
          example: 59.98
        currency:
          type: string
          example: USD
        status:
          type: string
          enum: [INITIATED, SUCCESS, FAILED, REFUNDED]
          example: SUCCESS

    OrderStatusHistory:
      type: object
      properties:
        id:
          type: integer
          example: 1
        fromStatus:
          type: string
          enum: [PENDING, PAID, SHIPPED, DELIVERED, CANCELLED]
          example: PENDING
        toStatus:
          type: string
          enum: [PENDING, PAID, SHIPPED, DELIVERED, CANCELLED]
          example: PAID
        changedBy:
          type: string
          enum: [SYSTEM, USER, ADMIN]
          example: SYSTEM
        createdAt:
          type: string
          format: date-time
          example: '2026-01-05T10:00:00Z'

    Order:
      type: object
      properties:
        id:
          type: integer
          example: 1
        userId:
          type: integer
          nullable: true
          example: 1
        addressId:
          type: integer
          nullable: true
          example: 3
        guestEmail:
          type: string
          nullable: true
          example: guest@example.com
        status:
          type: string
          enum: [PENDING, PAID, SHIPPED, DELIVERED, CANCELLED]
          example: PAID
        paymentStatus:
          type: string
          enum: [PENDING, PAID, FAILED, REFUNDED]
          example: PAID
        paymentMethod:
          type: string
          nullable: true
          enum: [STRIPE, PAYPAL, COD, MOCK]
          example: STRIPE
        currency:
          type: string
          example: USD
        totalPrice:
          type: number
          format: decimal
          example: 59.98
        shippingFee:
          type: number
          format: decimal
          example: 5
        taxAmount:
          type: number
          format: decimal
          example: 2.5
        createdAt:
          type: string
          format: date-time
          example: '2024-01-15T10:30:00Z'
        shippedAt:
          type: string
          format: date-time
          nullable: true
          example: '2024-01-16T10:30:00Z'
        deliveredAt:
          type: string
          format: date-time
          nullable: true
          example: '2024-01-18T10:30:00Z'
        user:
          $ref: '#/components/schemas/User'
        address:
          $ref: '#/components/schemas/Address'
        payment:
          $ref: '#/components/schemas/Payment'
        statusHistory:
          type: array
          items:
            $ref: '#/components/schemas/OrderStatusHistory'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'

    OrderResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Order'

    OrdersResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Order'

    CreateOrderRequest:
      type: object
      description: Provide either address_id or shipping_address
      required:
        - items
        - payment_placeholder
        - disclaimer_accepted
      properties:
        guest_email:
          type: string
          format: email
          description: Required for guest orders
          example: guest@example.com
        address_id:
          type: integer
          description: Use an existing address ID
          example: 3
        shipping_address:
          type: object
          description: Provide a structured address when no address_id is supplied
          required:
            - label
            - full_name
            - phone
            - street_line_1
            - city
            - state
            - postal_code
            - country
          properties:
            label:
              type: string
              example: Shipping
            full_name:
              type: string
              example: Jordan Lee
            phone:
              type: string
              example: "+15550000002"
            street_line_1:
              type: string
              example: 123 Wellness Ave
            street_line_2:
              type: string
              nullable: true
              example: Apt 4B
            city:
              type: string
              example: Austin
            state:
              type: string
              example: TX
            postal_code:
              type: string
              example: 73301
            country:
              type: string
              example: US
            is_default:
              type: boolean
              example: false
        items:
          type: array
          description: Array of product items in the order
          items:
            type: object
            required:
              - product_id
              - quantity
            properties:
              product_id:
                type: integer
                example: 1
              product_variant_id:
                type: integer
                nullable: true
                example: 10
              quantity:
                type: integer
                example: 2
        payment_placeholder:
          type: string
          description: Simulated payment information
          example: Card ending in 4242
        disclaimer_accepted:
          type: boolean
          description: Must acknowledge health disclaimer
          example: true

    Error:
      type: object
      properties:
        message:
          type: string
          description: Error message describing what went wrong
          example: Invalid credentials
        error:
          type: string
          description: Detailed error information (optional)