# src/backend/src/api/auth.ts
import { Router, Request, Response } from 'express';
import { userService } from '../services/userService';
import { auth } from '../lib/auth';

const router = Router();

// POST /api/auth/register
router.post('/register', async (req: Request, res: Response) => {
  try {
    const { email, password } = req.body;

    if (!email || !password) {
      return res.status(400).json({ message: 'Email and password are required' });
    }

    const existingUser = await userService.findUserByEmail(email);
    if (existingUser) {
      return res.status(409).json({ message: 'User already exists' });
    }

    const passwordHash = await auth.hashPassword(password);
    const { user } = await userService.createUserWithIdentity(
      email,
      req.body.firstName || 'User',
      req.body.lastName || 'User',
      'EMAIL',
      email,
      passwordHash
    );

    const token = auth.generateToken({ userId: user.id, email: user.email, role: user.role });

    res.status(201).json({
      token,
      user: {
        id: user.id,
        email: user.email,
        role: user.role,
      },
    });
  } catch (error) {
    console.error('Registration error:', error);
    res.status(500).json({ message: 'Registration failed' });
  }
});

// POST /api/auth/login
router.post('/login', async (req: Request, res: Response) => {
  try {
    const { email, password } = req.body;

    if (!email || !password) {
      return res.status(400).json({ message: 'Email and password are required' });
    }

    const identity = await userService.findIdentity('EMAIL', email);
    
    if (!identity || !identity.passwordHash) {
       // User not found or no password set for this identity
       return res.status(401).json({ message: 'Invalid credentials' });
    }

    const isValid = await auth.comparePassword(password, identity.passwordHash);
    if (!isValid) {
      return res.status(401).json({ message: 'Invalid credentials' });
    }

    const user = identity.user;
    const token = auth.generateToken({ userId: user.id, email: user.email, role: user.role });

    res.json({
      token,
      user: {
        id: user.id,
        email: user.email,
        role: user.role,
      },
    });

  } catch (error) {
    console.error('Login error:', error);
    res.status(500).json({ message: 'Login failed' });
  }
});

export default router;